Get the results for comparable recent property sales for a given property with additional property data

For a given address, extract property information including building data and Zestimates. At least one of zipcode or city/state information must be included.

GetDeepComps(zpid, count = 10, rentzestimate = FALSE, api_key,
  raw = FALSE)

Arguments

zpid
The Zillow property id

count
(integer) How many comparables to return?

rentzestimate
if TRUE, gets the rent zestimate.

api_key
character string specifying Zillow API key

raw
logical, if TRUE the raw XML data from the API call is returned (i.e., the original ZillowR call)

Value

If raw=T, a raw XML document. If raw=F (default), a data frame with columns corresponding to address information, Zestimates, and property information. The number of columns varies by property use type.

Examples

set_zillow_web_service_id('X1-ZWz181enkd4cgb_82rpe')
zapi_key = getOption('ZillowR-zws_id')
zpid='1341571'
GetDeepComps(zpid, count=10, rentzestimate=FALSE, api_key=zapi_key,raw=FALSE)

#>                address zipcode   city state      lat      long region_name
#> 1      733 Normandy Ct   67114 Newton    KS 38.06469 -97.35825      Newton
#> 2    906 E Broadway St   67114 Newton    KS 38.05029 -97.33022      Newton
#> 3        2137 Beltline   67114 Newton    KS 38.01605 -97.35733      Newton
#> 4      400 Homewood Ln   67114 Newton    KS 38.02281 -97.35082      Newton
#> 5   411 Meadowbrook Dr   67114 Newton    KS 38.02329 -97.35186      Newton
#> 6    132 Wheatridge Ct   67114 Newton    KS 38.02209 -97.34761      Newton
#> 7        7 S Kansas Ct   67114 Newton    KS 38.03397 -97.33806      Newton
#> 8  721 Country Club Dr   67114 Newton    KS 38.03479 -97.33421      Newton
#> 9      911 S Walnut St   67114 Newton    KS 38.03432 -97.33924      Newton
#> 10       921 S Plum St   67114 Newton    KS 38.03305 -97.34982      Newton
#> 11       500 Cherry Ln   67114 Newton    KS 38.03401 -97.35198      Newton
#>    region_id type zestimate zest_lastupdated zest_monthlychange zest_percentile
#> 1      19619 city    398794       2019-04-20                 NA              99
#> 2      19619 city     90372       2019-04-20                 NA              50
#> 3      19619 city    269883       2019-04-20                 NA              98
#> 4      19619 city    176742       2019-04-20                 NA              88
#> 5      19619 city    167591       2019-04-20                 NA              87
#> 6      19619 city    148759       2019-04-20                 NA              80
#> 7      19619 city    203142       2019-04-20                 NA              94
#> 8      19619 city    232841       2019-04-20                 NA              96
#> 9      19619 city    115006       2019-04-20                 NA              63
#> 10     19619 city    114302       2019-04-20                 NA              62
#> 11     19619 city    125739       2019-04-20                 NA              70
#>    zestimate_low zestimate_high      zpid compscore bathrooms bedrooms
#> 1         354927         438673   1341571        NA       3.5        7
#> 2          85853          98505   1339410        16       2.0        5
#> 3         256389         283377 109818062        18       3.0        6
#> 4         167905         185579  80865410        17       3.0        5
#> 5         157536         182674  80865235        17       3.0        4
#> 6         141321         156197   1341819        18       3.0        5
#> 7         192985         213299 109780123        15       3.0        3
#> 8         221199         244483   1340091        16       3.0        3
#> 9         109256         120756   1338699        17       2.0        3
#> 10        108587         120017   1338444        18       2.0        3
#> 11        116937         132026   1338374        15       3.0        3
#>    finishedSqFt lastSoldDate lastSoldPrice lotSizeSqFt taxAssessment
#> 1          3229   06/01/2007        429000       17172        440800
#> 2          1726   12/18/2018            NA        7840        110500
#> 3          2669   06/15/2018        262900       13503        228000
#> 4          2464   03/08/2019            NA        8276        158000
#> 5          2700   05/31/2018            NA       13068        131580
#> 6          1764   08/10/2018            NA       10018        126680
#> 7          2492   05/25/2018            NA       17424        186100
#> 8          2805   09/25/2018            NA       23522        217350
#> 9          2371   05/25/2018            NA       13939         96900
#> 10         1920   08/10/2018            NA        7405         88500
#> 11         2100   07/20/2018            NA        8800         92920
#>    taxAssessmentYear totalRooms yearBuilt
#> 1               2018         13      1994
#> 2               2018          7      1926
#> 3               2018         18      2008
#> 4               2018          8      2006
#> 5               2018          7      2006
#> 6               2018          8      1999
#> 7               2018          6      2002
#> 8               2018          7      1979
#> 9               2018          6      1965
#> 10              2018         11      1958
#> 11              2018         10      1975