Get the 'Deep Search' results from the Zillow API

For a given address, extract property information including building data and Zestimates. At least one of zipcode or city/state information must be included.

GetDeepSearchResults(address, city = NULL, state = NULL,
  zipcode = NULL, rentzestimate = F, api_key, raw = F)

Arguments

address
(required) A character string specifying the address

city
An optional character string specifying the city name

state
An optional character string specifying the state name (abbreviated length 2)

zipcode
An optional character or numeric string of length 5 specifying the zip code

rentzestimate
if TRUE, gets the rent zestimate.

api_key
(required) A unique character string specifying the Zillow user's API key

raw
(optional) If TRUE, the raw xml document from the API call is returned. If FALSE, the data is extracted and formatted into a data frame.

Value

If raw=T, the XML document directly from the API call. If raw=F (default), a data frame with columns corresponding to address information, Zestimates, and property information. The number of columns varies by property use type.

Examples

# NOT RUN {
zapi_key = getOption('ZillowR-zws_id')
# }

# NOT RUN {
GetDeepSearchResults(address='600 S. Quail Ct.', zipcode='67114',
 rentzestimate=FALSE, api_key=zapi_key)

# }

# NOT RUN {
GetDeepSearchResults(address='312 Hayward Ave.', city='Ames', state='IA',
 rentzestimate=TRUE, zipcode='50014', api_key=zapi_key)
# }