Get the 'Deep Search' results from the Zillow API

A wrapper for GetDeepSearchResults that takes data frames with address information as inputs rather than a single address, and returns the search results for all addresses

GetDeepSearchResults_dataframe(.df, col.address, col.zipcode = NULL,
  col.city = NULL, col.state = NULL, rentzestimate = F, api_key,
  raw = FALSE)

Arguments

.df
a dataframe with address data

col.address
a numeric value specifying which column in .df contains the addresses

col.zipcode
a numeric value specifying which column contains the zipcodes (required if no city/state information)

col.city
a numeric value specifying which column contains the city (required if no zipcode information)

col.state
a numeric value specifying which column contains the state abbrev. (required if no zipcode information)

rentzestimate
if TRUE, additional zestimates for rental prices are returned.

api_key
a character string specifying the unique Zillow API key

raw
logical, if TRUE, a list with raw XML documents (i.e., the original ZillowR call) for each property in the data frame is returned

Value

A data frame with columns corresponding to address, property, and Zestimate information

Examples

library(tidyverse)

zapi_key = 'X1-ZWz181ckl6u9e3_9k5bc'

adrs <- c('733 Normandy Ct', '600 South Quail Ct','2902 Wood St', '3425 Locust St.')
data.frame(address=adrs,zipcode=c('67114','67114','50014', '64109'),
state=c("KS","KS","IA","MO"), city=c("Newton","Newton","Ames","Kansas City")) %>%
GetDeepSearchResults_dataframe(col.address=1, col.zipcode=2, col.city=4,
col.state=3,api_key=zapi_key)

#>                   address zipcode        city state      lat      long
#> 1         733 Normandy Ct   67114      Newton    KS 38.06469 -97.35825
#> 2          600 S Quail Ct   67114      Newton    KS 38.01740 -97.33380
#> 3            2902 Wood St   50014        Ames    IA 42.02181 -93.65714
#> 4  3425 Locust St APT 107   64109 Kansas City    MO 39.06428 -94.57948
#> 5  3425 Locust St APT 104   64109 Kansas City    MO 39.06428 -94.57948
#> 6  3425 Locust St APT 206   64109 Kansas City    MO 39.06428 -94.57948
#> 7  3425 Locust St APT 103   64109 Kansas City    MO 39.06428 -94.57948
#> 8  3425 Locust St APT 203   64109 Kansas City    MO 39.06428 -94.57948
#> 9  3425 Locust St APT 105   64109 Kansas City    MO 39.06428 -94.57948
#> 10 3425 Locust St APT 205   64109 Kansas City    MO 39.06428 -94.57948
#> 11 3425 Locust St APT 102   64109 Kansas City    MO 39.06428 -94.57948
#>               region_name region_id         type zestimate zest_lastupdated
#> 1                  Newton     19619         city    398794       2019-04-20
#> 2                  Newton     19619         city    231324       2019-04-20
#> 3  College Creek/Old Ames    764177 neighborhood    150984       2019-04-20
#> 4         North Hyde Park    274720 neighborhood        NA       1970-01-01
#> 5         North Hyde Park    274720 neighborhood        NA       1970-01-01
#> 6         North Hyde Park    274720 neighborhood        NA       1970-01-01
#> 7         North Hyde Park    274720 neighborhood        NA       1970-01-01
#> 8         North Hyde Park    274720 neighborhood        NA       1970-01-01
#> 9         North Hyde Park    274720 neighborhood        NA       1970-01-01
#> 10        North Hyde Park    274720 neighborhood        NA       1970-01-01
#> 11        North Hyde Park    274720 neighborhood        NA       1970-01-01
#>    zest_monthlychange zest_percentile zestimate_low zestimate_high       zpid
#> 1                  NA               0        354927         438673    1341571
#> 2                  NA               0        210505         245203    1340244
#> 3                  NA               0        143435         158533   93961896
#> 4                  NA               0            NA             NA 2088988714
#> 5                  NA               0            NA             NA 2090030184
#> 6                  NA               0            NA             NA 2091286972
#> 7                  NA               0            NA             NA 2092556040
#> 8                  NA               0            NA             NA 2093235689
#> 9                  NA               0            NA             NA 2094825592
#> 10                 NA               0            NA             NA 2095110197
#> 11                 NA               0            NA             NA 2095113899
#>    bathrooms bedrooms finishedSqFt lastSoldDate lastSoldPrice lotSizeSqFt
#> 1        3.5        7         3229   06/01/2007        429000       17172
#> 2        3.0        4         1726   06/01/2007        230000       33024
#> 3        1.0        3         1298   01/19/2015        131000        8212
#> 4        1.0        1          450         <NA>            NA          NA
#> 5        1.0        1          450         <NA>            NA          NA
#> 6        1.0        1          450         <NA>            NA          NA
#> 7        1.0        1          450         <NA>            NA          NA
#> 8        1.0        1          450         <NA>            NA          NA
#> 9        1.0        1          450         <NA>            NA          NA
#> 10       1.0        1          450         <NA>            NA          NA
#> 11       1.0        1          450         <NA>            NA          NA
#>    taxAssessment taxAssessmentYear totalRooms yearBuilt
#> 1         440800              2018         13      1994
#> 2         214800              2018          8      1993
#> 3         138800              2018          6      1941
#> 4             NA                NA         NA        NA
#> 5             NA                NA         NA        NA
#> 6             NA                NA         NA        NA
#> 7             NA                NA         NA        NA
#> 8             NA                NA         NA        NA
#> 9             NA                NA         NA        NA
#> 10            NA                NA         NA        NA
#> 11            NA                NA         NA        NA