Get Zillow Home Value Index time series data

Reads the static .csv file for Zillow Home Values series by type and geography, building a URL path to the .csv file. Options are available for a variety of home sizes and types, as well as tiers if you wish to view other data that is aggregated.

get_ZHVI_series(bedrooms = 1, geography = "Metro", allhomes = F,
  tier = "ALL", summary = F, other = NULL)

Arguments

bedrooms
a numeric value specifying the number of bedrooms. If not needed, leave at the default (1)

geography
(required) string specifying the desired geographic region to summarise. Choices are 'Metro', 'City', 'State', 'Neighborhood', 'Zip', and 'County'.

allhomes
logical, set to TRUE If you do not want a specific # of bedrooms or type of home

tier
one of 'ALL','T' (for top), 'B' (for bottom), or 'M' (for median)

summary
logical, if TRUE, the ZHVI summary for all homes is returned

other
Other possible options from the Zillow home data series. Default is NULL. Possible options are strings "Median Home Price Per Sq Ft", and "Increasing" or "Decreasing"

Value

A tibble. Columns returned correspond to geographic identification information and dates for which the time series observations are available.

Examples

#All homes, bottom tier, by zipcode
get_ZHVI_series(allhomes=TRUE, tier="B",geography="Zip")

#> [1] "attempting to read file:"
#> [1] "http://files.zillowstatic.com/research/public/Zip/Zip_Zhvi_BottomTier.csv"

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   RegionName = col_character(),
#>   City = col_character(),
#>   State = col_character(),
#>   Metro = col_character(),
#>   CountyName = col_character()
#> )

#> See spec(...) for full column specifications.

#> # A tibble: 12,996 x 283
#>    RegionID RegionName City  State Metro CountyName SizeRank `1996-04` `1996-05`
#>       <dbl> <chr>      <chr> <chr> <chr> <chr>         <dbl>     <dbl>     <dbl>
#>  1    61639 10025      New ~ NY    New ~ New York ~        1    106900    106500
#>  2    84654 60657      Chic~ IL    Chic~ Cook Coun~        2     75900     76500
#>  3    61637 10023      New ~ NY    New ~ New York ~        3    188600    188800
#>  4    91982 77494      Katy  TX    Hous~ Harris Co~        4    137300    136700
#>  5    84616 60614      Chic~ IL    Chic~ Cook Coun~        5     84400     85600
#>  6    91940 77449      Katy  TX    Hous~ Harris Co~        6     76800     76700
#>  7    93144 79936      El P~ TX    El P~ El Paso C~        7     65800     65800
#>  8    91733 77084      Hous~ TX    Hous~ Harris Co~        8     75800     75800
#>  9    61616 10002      New ~ NY    New ~ New York ~        9    132800    132900
#> 10    61807 10467      New ~ NY    New ~ Bronx Cou~       10        NA        NA
#> # ... with 12,986 more rows, and 274 more variables: `1996-06` <dbl>,
#> #   `1996-07` <dbl>, `1996-08` <dbl>, `1996-09` <dbl>, `1996-10` <dbl>,
#> #   `1996-11` <dbl>, `1996-12` <dbl>, `1997-01` <dbl>, `1997-02` <dbl>,
#> #   `1997-03` <dbl>, `1997-04` <dbl>, `1997-05` <dbl>, `1997-06` <dbl>,
#> #   `1997-07` <dbl>, `1997-08` <dbl>, `1997-09` <dbl>, `1997-10` <dbl>,
#> #   `1997-11` <dbl>, `1997-12` <dbl>, `1998-01` <dbl>, `1998-02` <dbl>,
#> #   `1998-03` <dbl>, `1998-04` <dbl>, `1998-05` <dbl>, `1998-06` <dbl>,
#> #   `1998-07` <dbl>, `1998-08` <dbl>, `1998-09` <dbl>, `1998-10` <dbl>,
#> #   `1998-11` <dbl>, `1998-12` <dbl>, `1999-01` <dbl>, `1999-02` <dbl>,
#> #   `1999-03` <dbl>, `1999-04` <dbl>, `1999-05` <dbl>, `1999-06` <dbl>,
#> #   `1999-07` <dbl>, `1999-08` <dbl>, `1999-09` <dbl>, `1999-10` <dbl>,
#> #   `1999-11` <dbl>, `1999-12` <dbl>, `2000-01` <dbl>, `2000-02` <dbl>,
#> #   `2000-03` <dbl>, `2000-04` <dbl>, `2000-05` <dbl>, `2000-06` <dbl>,
#> #   `2000-07` <dbl>, `2000-08` <dbl>, `2000-09` <dbl>, `2000-10` <dbl>,
#> #   `2000-11` <dbl>, `2000-12` <dbl>, `2001-01` <dbl>, `2001-02` <dbl>,
#> #   `2001-03` <dbl>, `2001-04` <dbl>, `2001-05` <dbl>, `2001-06` <dbl>,
#> #   `2001-07` <dbl>, `2001-08` <dbl>, `2001-09` <dbl>, `2001-10` <dbl>,
#> #   `2001-11` <dbl>, `2001-12` <dbl>, `2002-01` <dbl>, `2002-02` <dbl>,
#> #   `2002-03` <dbl>, `2002-04` <dbl>, `2002-05` <dbl>, `2002-06` <dbl>,
#> #   `2002-07` <dbl>, `2002-08` <dbl>, `2002-09` <dbl>, `2002-10` <dbl>,
#> #   `2002-11` <dbl>, `2002-12` <dbl>, `2003-01` <dbl>, `2003-02` <dbl>,
#> #   `2003-03` <dbl>, `2003-04` <dbl>, `2003-05` <dbl>, `2003-06` <dbl>,
#> #   `2003-07` <dbl>, `2003-08` <dbl>, `2003-09` <dbl>, `2003-10` <dbl>,
#> #   `2003-11` <dbl>, `2003-12` <dbl>, `2004-01` <dbl>, `2004-02` <dbl>,
#> #   `2004-03` <dbl>, `2004-04` <dbl>, `2004-05` <dbl>, `2004-06` <dbl>,
#> #   `2004-07` <dbl>, `2004-08` <dbl>, `2004-09` <dbl>, ...

#2 bedrooms by city
get_ZHVI_series(bedrooms=2,geography="City")

#> [1] "attempting to read file:"
#> [1] "http://files.zillowstatic.com/research/public/City/City_Zhvi_2bedroom.csv"

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   RegionName = col_character(),
#>   State = col_character(),
#>   Metro = col_character(),
#>   CountyName = col_character()
#> )

#> See spec(...) for full column specifications.

#> # A tibble: 6,980 x 282
#>    RegionID RegionName State Metro CountyName SizeRank `1996-04` `1996-05`
#>       <dbl> <chr>      <chr> <chr> <chr>         <dbl>     <dbl>     <dbl>
#>  1    12447 Los Angel~ CA    Los ~ Los Angel~        1    130700    130500
#>  2    39051 Houston    TX    Hous~ Harris Co~        2     50300     50200
#>  3    17426 Chicago    IL    Chic~ Cook Coun~        3     89700     90300
#>  4     6915 San Anton~ TX    San ~ Bexar Cou~        4     48300     48400
#>  5    13271 Philadelp~ PA    Phil~ Philadelp~        5     42300     42100
#>  6    18959 Las Vegas  NV    Las ~ Clark Cou~        6     90400     90500
#>  7    54296 San Diego  CA    San ~ San Diego~        7    127800    127400
#>  8    38128 Dallas     TX    Dall~ Dallas Co~        8     54600     54700
#>  9    10221 Austin     TX    Aust~ Travis Co~        9    135800    135200
#> 10    33839 San Jose   CA    San ~ Santa Cla~       10    168000    167900
#> # ... with 6,970 more rows, and 274 more variables: `1996-06` <dbl>,
#> #   `1996-07` <dbl>, `1996-08` <dbl>, `1996-09` <dbl>, `1996-10` <dbl>,
#> #   `1996-11` <dbl>, `1996-12` <dbl>, `1997-01` <dbl>, `1997-02` <dbl>,
#> #   `1997-03` <dbl>, `1997-04` <dbl>, `1997-05` <dbl>, `1997-06` <dbl>,
#> #   `1997-07` <dbl>, `1997-08` <dbl>, `1997-09` <dbl>, `1997-10` <dbl>,
#> #   `1997-11` <dbl>, `1997-12` <dbl>, `1998-01` <dbl>, `1998-02` <dbl>,
#> #   `1998-03` <dbl>, `1998-04` <dbl>, `1998-05` <dbl>, `1998-06` <dbl>,
#> #   `1998-07` <dbl>, `1998-08` <dbl>, `1998-09` <dbl>, `1998-10` <dbl>,
#> #   `1998-11` <dbl>, `1998-12` <dbl>, `1999-01` <dbl>, `1999-02` <dbl>,
#> #   `1999-03` <dbl>, `1999-04` <dbl>, `1999-05` <dbl>, `1999-06` <dbl>,
#> #   `1999-07` <dbl>, `1999-08` <dbl>, `1999-09` <dbl>, `1999-10` <dbl>,
#> #   `1999-11` <dbl>, `1999-12` <dbl>, `2000-01` <dbl>, `2000-02` <dbl>,
#> #   `2000-03` <dbl>, `2000-04` <dbl>, `2000-05` <dbl>, `2000-06` <dbl>,
#> #   `2000-07` <dbl>, `2000-08` <dbl>, `2000-09` <dbl>, `2000-10` <dbl>,
#> #   `2000-11` <dbl>, `2000-12` <dbl>, `2001-01` <dbl>, `2001-02` <dbl>,
#> #   `2001-03` <dbl>, `2001-04` <dbl>, `2001-05` <dbl>, `2001-06` <dbl>,
#> #   `2001-07` <dbl>, `2001-08` <dbl>, `2001-09` <dbl>, `2001-10` <dbl>,
#> #   `2001-11` <dbl>, `2001-12` <dbl>, `2002-01` <dbl>, `2002-02` <dbl>,
#> #   `2002-03` <dbl>, `2002-04` <dbl>, `2002-05` <dbl>, `2002-06` <dbl>,
#> #   `2002-07` <dbl>, `2002-08` <dbl>, `2002-09` <dbl>, `2002-10` <dbl>,
#> #   `2002-11` <dbl>, `2002-12` <dbl>, `2003-01` <dbl>, `2003-02` <dbl>,
#> #   `2003-03` <dbl>, `2003-04` <dbl>, `2003-05` <dbl>, `2003-06` <dbl>,
#> #   `2003-07` <dbl>, `2003-08` <dbl>, `2003-09` <dbl>, `2003-10` <dbl>,
#> #   `2003-11` <dbl>, `2003-12` <dbl>, `2004-01` <dbl>, `2004-02` <dbl>,
#> #   `2004-03` <dbl>, `2004-04` <dbl>, `2004-05` <dbl>, `2004-06` <dbl>,
#> #   `2004-07` <dbl>, `2004-08` <dbl>, `2004-09` <dbl>, ...

#the ZHVI summary for all homes by State
get_ZHVI_series(geography="State", summary=TRUE)

#> [1] "attempting to read file:"
#> [1] "http://files.zillowstatic.com/research/public/State/State_Zhvi_Summary_AllHomes.csv"

#> Parsed with column specification:
#> cols(
#>   Date = col_date(format = ""),
#>   RegionID = col_double(),
#>   RegionName = col_character(),
#>   SizeRank = col_double(),
#>   Zhvi = col_double(),
#>   MoM = col_double(),
#>   QoQ = col_double(),
#>   YoY = col_double(),
#>   `5Year` = col_double(),
#>   `10Year` = col_double(),
#>   PeakMonth = col_character(),
#>   PeakQuarter = col_character(),
#>   PeakZHVI = col_double(),
#>   PctFallFromPeak = col_double(),
#>   LastTimeAtCurrZHVI = col_character()
#> )

#> # A tibble: 52 x 15
#>    Date       RegionID RegionName SizeRank   Zhvi      MoM      QoQ    YoY
#>    <date>        <dbl> <chr>         <dbl>  <dbl>    <dbl>    <dbl>  <dbl>
#>  1 2019-03-31   102001 United St~        0 226700  1.77e-3  0.00980 0.0663
#>  2 2019-03-31        9 California        1 548700 -2.18e-3 -0.00146 0.0258
#>  3 2019-03-31       54 Texas             2 197300  4.58e-3  0.0181  0.0787
#>  4 2019-03-31       43 New York          3 296800  7.47e-3  0.0238  0.0574
#>  5 2019-03-31       14 Florida           4 233700  8.57e-4  0.00950 0.0628
#>  6 2019-03-31       21 Illinois          5 181100  2.21e-3  0.00948 0.0414
#>  7 2019-03-31       47 Pennsylva~        6 173900  1.73e-3  0.00637 0.0463
#>  8 2019-03-31       44 Ohio              7 139600  2.15e-3  0.00867 0.0697
#>  9 2019-03-31       30 Michigan          8 151700  3.31e-3  0.0120  0.0851
#> 10 2019-03-31       16 Georgia           9 188500  2.66e-3  0.0151  0.102 
#> # ... with 42 more rows, and 7 more variables: `5Year` <dbl>, `10Year` <dbl>,
#> #   PeakMonth <chr>, PeakQuarter <chr>, PeakZHVI <dbl>, PctFallFromPeak <dbl>,
#> #   LastTimeAtCurrZHVI <chr>