Data on 51 comparable homes in the Newton, KS, USA area

A dataset containing property attributes and value information for 210 homes in the Newton, KS area. This dataset was generated by chaining together the `GetDeepSearchResults` and `GetDeepComps` functions (as shown in the package vignette) to find comparable homes starting from the initial single address of 600 S. Quail Ct.

newtonhouses

Format

A data frame with 51 rows and 26 variables:

address

Property street address

zipcode

Five digit zipcode

city

City

state

State abbreviation

lat

Latitude of location

long

Longitude of location

region_name

region_id

Zillow-assigned five digit region ID

type

Type of region (i.e., Neighborhood, City,...)

zestimate

Zillow's estimate of the home value

zest_lastupdated

The date of the last update to the Zestimate

zest_monthlychange

Zestimate change from previous month

zest_percentile

Percentile rank of the property Zestimate within the region

zestimate_low

The lower bound of the range of valuation

zestimate_high

The upper bound of the range of valuation

zpid

The unique Zillow-assigned property ID number

bathrooms

Number of bathrooms in the property

bedrooms

Number of bedrooms in the property

finishedSqFt

Size of finished area in the property, in square feet

lastSoldDate

Date the property was last sold (if applicable)

lastSoldPrice

The most recent sale price of the property (if applicable)

lotSizeSqFt

Size of the lot the property is on, in square feet

taxAssessment

The most recent available local property tax assessment value, in dollars

taxAssessmentYear

Year corresponding to the most recent property tax assessment

totalRooms

Total number of rooms in the property

yearBuilt

Year the property was built

...

Source

http://www.zillow.com/